refactor(structure): Organize components into 'previews' subfolder
Move preview components (AboutPreview, ContactPreview, ProjectsPreview, SkillsPreview) into a new `components/previews` directory. This improves project organization by grouping related preview-specific components together.
